+++ src/app/ui/widget/AdvancedPageContent.tsx
@@ -2,8 +2,8 @@
 
 import { useEffect, useState } from 'react';
 import { WidgetEvent, useWidgetEvents } from '@lifi/widget';
-import { OrdersSection } from '@/components/LimitOrders/OrdersSection';
-import { MarketPriceSection } from '@/components/LimitOrders/MarketPriceSection';
+import { OrdersSection } from '@/components/LimitOrders/OrdersSection/OrdersSection';
+import { MarketPriceSection } from '@/components/LimitOrders/MarketPriceSection/MarketPriceSection';
 import { useWidgetSidePanelStore } from '@/stores/widgetSidePanel/WidgetSidePanelStore';
 import { MainWidgetPageContent } from './MainWidgetPageContent';
 import { WidgetSidePanel } from './WidgetSidePanel';
@@ -11,8 +11,10 @@ import CloseFullscreenIcon from '@mui/icons-material/CloseFullscreen';
 import OpenInFullIcon from '@mui/icons-material/OpenInFull';
 import { IconButton } from '@/components/core/buttons/IconButton/IconButton';
 import { Size, Variant } from '@/components/core/buttons/types';
+import { useTranslation } from 'react-i18next';
 
 export const AdvancedPageContent = () => {
+  const { t } = useTranslation();
   const widgetEvents = useWidgetEvents();
   const [isLimitTabActive, setIsLimitTabActive] = useState(false);
   const isSidePanelExpanded = useWidgetSidePanelStore(
@@ -41,10 +43,14 @@ export const AdvancedPageContent = () => {
   const expandButton = (
     <IconButton
       size={Size.SM}
-      variant={Variant.AlphaLight}
+      variant={Variant.AlphaDark}
       onClick={toggleSidePanelExpanded}
       aria-expanded={isSidePanelExpanded}
-      aria-label={isSidePanelExpanded ? 'Collapse panels' : 'Expand panels'}
+      aria-label={
+        isSidePanelExpanded
+          ? t('limitOrders.table.actions.collapsePanels')
+          : t('limitOrders.table.actions.expandPanels')
+      }
       data-testid="orders-section-expand"
       sx={{ height: 32, width: 32 }}
     >
