+++ src/components/Widgets/Widget.tsx
@@ -41,7 +41,6 @@ export function Widget({
   autoHeight,
   isLoading,
   disableTabNavigation = false,
-  useLayoutWelcomeScreenMargins,
 }: WidgetProps) {
   const [configTheme] = useThemeStore((state) => [state.configTheme]);
   const formRef = useRef<FormState>(null);
@@ -183,7 +182,6 @@ export function Widget({
       welcomeScreenClosed={welcomeScreenClosed || !enabled}
       autoHeight={autoHeight}
       contributionDisplayed={contributionDisplayed}
-      useLayoutWelcomeScreenMargins={useLayoutWelcomeScreenMargins}
     >
       <BaseWidget
         type="main"
