+++ src/hooks/useMainPaths.ts
@@ -15,10 +15,11 @@ export const useMainPaths = (): useMainPathsProps => {
   const isGas = matchesPath(AppPaths.Gas);
   const isBuy = matchesPath(AppPaths.Buy);
   const isPrivate = matchesPath(AppPaths.Private);
+  const isAdvanced = matchesPath(AppPaths.Advanced);
   const isExchange = pathname === AppPaths.Main;
 
   return {
-    isMainPaths: isGas || isBuy || isPrivate || isExchange,
+    isMainPaths: isGas || isBuy || isPrivate || isExchange || isAdvanced,
     isExchange,
   };
 };
